+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 changelog
 
+<a name="1.68.1"></a>
+## [1.68.1](https://github.com/33cn/plugin/compare/v1.68.0...v1.68.1) (2023-03-14)
+
+
+### Bug Fixes
+
+* fix rpc config for parachain ([cd9ad35](https://github.com/33cn/plugin/commit/cd9ad35))
+
+<a name="1.68.0"></a>
+# [1.68.0](https://github.com/33cn/plugin/compare/v1.67.6...v1.68.0) (2023-02-21)
+
+
+### Features
+
+* add consensus commiter plugin rollup(33cn/chain33#1268) ([b0aca7d](https://github.com/33cn/plugin/commit/b0aca7d)), closes [33cn/chain33#1268](https://github.com/33cn/chain33/issues/1268)
+
+<a name="1.67.6"></a>
+## [1.67.6](https://github.com/33cn/plugin/compare/v1.67.5...v1.67.6) (2023-01-31)
+
+
+### Bug Fixes
+
+* add return for get evm tx recevier ([533f35a](https://github.com/33cn/plugin/commit/533f35a))
+
+<a name="1.67.5"></a>
+## [1.67.5](https://github.com/33cn/plugin/compare/v1.67.4...v1.67.5) (2023-01-16)
+
+
+### Bug Fixes
+
+* add evm mix address fork ([7fecaf4](https://github.com/33cn/plugin/commit/7fecaf4))
+
 <a name="1.67.4"></a>
 ## [1.67.4](https://github.com/33cn/plugin/compare/v1.67.3...v1.67.4) (2022-10-11)